exercicios/arduino and python/requirements.txt
Pillow==5.3.0
pyserial==3.4
